Moved all custom validation checks to a separate .groovy file.
